# Conflicts:
#	Training2020/Question banks system/ClientQuestionBank/src/app/Components/Admin/Question/list/list.component.ts
#	Training2020/Question banks system/ClientQuestionBank/src/app/Services/question.service.ts
#	Training2020/Question banks system/ClientQuestionBank/src/app/Services/test.service.ts
#	Training2020/Question banks system/QuestionBankDB/QuestionBankDB/Services/QuestionService.cs

+++ Training2020/Question banks system/QuestionBankDB/QuestionBankDB/Models/Question.cs
@@ -26,13 +26,13 @@ namespace QuestionBankDB.Models
         public string ThemeName { get; set; }
 
         [BsonElement("level")]
-        public long Level { get; set; }
+        public byte Level { get; set; }
 
         [BsonElement("point")]
-        public long Point { get; set; }
+        public byte Point { get; set; }
 
         [BsonElement("timeallow")]
-        public long Timeallow { get; set; }
+        public byte Timeallow { get; set; }
 
         [BsonElement("status")]
         public bool Status { get; set; }
